Create Borderlands_Counter1.1.ahk
Version 1.1

Simple counter like V1 but has added functionality to add name of item you're farming for to add to screen.
Created a seperate file for those who only want the run counter and those who want to add the item as well.

Creates GUI always on top with transparent background and disables Menu and Caption to make whole window transparent except text. Adds check to see if farming.txt is in default directory (Directory that .ahk is in), and creates it if not there.

Adds Bold basic text to count run number and dispay farmed item,  increments the run number after each press.

Customization:
- Change Size of text
- Change Color of text (in hex)
-Change postion of text on screen
-Type name of item in farming.txt first line and save for item to display.